feat: move usage metrics controller to system package
Moves the controller to the `system` package to group system-level APIs in one place,
as more endpoints are expected to be moved under the `/system` path in the future.
