+++ zeebe/gateway-rest/src/test/java/io/camunda/zeebe/gateway/rest/controller/system/UsageMetricsControllerTest.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.zeebe.gateway.rest.controller;
+package io.camunda.zeebe.gateway.rest.controller.system;
 
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.verify;
